Rename demangler.rs -> demangle.rs
